# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mait AI Platform API
  version: 1.0.0
  description: Artificial Insemination field operations platform for Shwetdhara Milk
    Producer Company. Serves the Mait mobile app, the admin web portal and the Indent
    Easy integration.
paths:
...
  /api/v1/admin/users/activate-mait/:
    post:
      operationId: admin_users_activate_mait_create
      description: |-
        Gives an existing SAP Sahayak record a login and sets the mobile number their OTP will be sent to.

        The number is set here because 2,886 of 3,110 Sahayak records arrive from SAP without one, and OTP is a Mait's only way in. No password is created — that is what keeps OTP the sole route into a field account.
      summary: Activate a Mait
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaitActivationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MaitActivationRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MaitActivationRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
...
  /api/v1/auth/otp/verify/:
    post:
      operationId: auth_otp_verify_create
      description: Expires after 5 minutes; 3 wrong attempts force a resend (SRS §6.5.1).
        Every send and every attempt is logged for fraud review.
      summary: Verify a login OTP
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerifyRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OTPVerifyRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OTPVerifyRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          description: ''
...
components:
  schemas:
    AdminUser:
      type: object
      description: Read shape for the user list (SRS §9.10).
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
        full_name:
          type: string
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        mobile_no:
          type: string
          readOnly: true
          description: Required for Maits — this is where the login OTP is sent.
        role:
          allOf:
          - $ref: '#/components/schemas/Role412Enum'
          readOnly: true
        role_display:
          type: string
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
        last_login_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        mait_name:
          type: string
          readOnly: true
        sahayak_vendor_code:
          type: string
          readOnly: true
        assigned_mpp_count:
          type: integer
          readOnly: true
      required:
      - assigned_mpp_count
      - created_at
      - email
      - full_name
      - id
      - is_active
      - last_login_at
      - mait_name
      - mobile_no
      - role
      - role_display
      - sahayak_vendor_code
      - username
...
    MaitActivationRequest:
      type: object
      description: |-
        Give an existing SAP Sahayak record a login (SRS §6.8.2).

        ``mobile_no`` is accepted here because 2,886 of 3,110 Sahayak records arrive from SAP
        with no number at all, and OTP is a Mait's only way in. Setting it at activation is what
        SRS §6.8.2 describes — "activated by Admin with a mobile number for OTP-based first
        login" — and without it those Maits could never sign in
        (docs/DATA_FINDINGS.md §1).
      properties:
        sahayak_vendor_code:
          type: string
          minLength: 1
        mobile_no:
          type: string
          minLength: 1
          pattern: ^[6-9]\d{9}$
        username:
          type: string
          minLength: 1
          description: Defaults to the Sahayak vendor code.
          maxLength: 64
      required:
      - mobile_no
      - sahayak_vendor_code
...
    OTPVerifyRequest:
      type: object
      description: Verify the login OTP and exchange it for tokens.
      properties:
        mobile_no:
          type: string
          minLength: 1
          pattern: ^[6-9]\d{9}$
        otp:
          type: string
          minLength: 4
          maxLength: 8
      required:
      - mobile_no
      - otp
...
    Role412Enum:
      enum:
      - super_admin
      - admin
      - mait
      type: string
      description: |-
        * `super_admin` - Super Admin
        * `admin` - Admin / Back-office
        * `mait` - Mait (Field Agent)
...
    TokenPair:
      type: object
      description: The JWT pair every successful login returns.
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
          readOnly: true
      required:
      - access
      - refresh
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...